Fix several overflows

+++ arch/SystemZ/SystemZDisassembler.c
@@ -41,7 +41,7 @@
 
 #define DEBUG_TYPE "systemz-disassembler"
 
-static DecodeStatus getInstruction(MCInst *Instr, uint64_t *Size, const uint8_t *Bytes,
+static DecodeStatus getInstruction(MCInst *Instr, uint16_t *Size, const uint8_t *Bytes,
 			    size_t BytesLen, uint64_t Address,
 			    SStream *CStream);
 
@@ -365,7 +365,7 @@ static DecodeStatus decodePC32DBLOperand(MCInst *Inst, uint64_t Imm,
 
 #include "SystemZGenDisassemblerTables.inc"
 
-static DecodeStatus getInstruction(MCInst *MI, uint64_t *Size, const uint8_t *Bytes,
+static DecodeStatus getInstruction(MCInst *MI, uint16_t *Size, const uint8_t *Bytes,
 			    size_t BytesLen, uint64_t Address, SStream *CS)
 {
 	// Get the first two bytes of the instruction.
@@ -402,7 +402,7 @@ static DecodeStatus getInstruction(MCInst *MI, uint64_t *Size, const uint8_t *By
 
 DecodeStatus SystemZ_LLVM_getInstruction(csh handle, const uint8_t *Bytes,
 				     size_t BytesLen, MCInst *MI,
-				     uint64_t *Size, uint64_t Address,
+				     uint16_t *Size, uint64_t Address,
 				     void *Info)
 {
 	return getInstruction(MI, Size, Bytes, BytesLen, MI->address, NULL);
